Rename package to jupyter_geotiler
