+++ .github/workflows/pr-rtd-link.yml
@@ -19,7 +19,7 @@ jobs:
     steps:
       - uses: "readthedocs/actions/preview@b8bba1484329bda1a3abe986df7ebc80a8950333" # v1
         with:
-          project-slug: "jupyter-xarray-tiler"
+          project-slug: "jupyter-geotiler"
           message-template: |
             ---
             :mag: Docs preview: {docs-pr-index-url}
