+++ .github/workflows/publish.yml
@@ -28,7 +28,7 @@ jobs:
     runs-on: "ubuntu-latest"
     environment:
       name: "pypi"
-      url: "https://pypi.org/p/jupyter-xarray-tiler"
+      url: "https://pypi.org/p/jupyter-geotiler"
     permissions:
       id-token: "write"
     steps:
