+++ .gitignore
@@ -9,9 +9,9 @@ node_modules/
 *.egg-info/
 .ipynb_checkpoints
 *.tsbuildinfo
-jupyter_xarray_tiler/labextension
+jupyter_geotiler/labextension
 # Version file is handled by hatchling
-jupyter_xarray_tiler/_version.py
+jupyter_geotiler/_version.py
 
 # Integration tests
 ui-tests/test-results/
