+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
 # Contributing
 
 Please see the
-[Contributor Guide in our documentation](https://jupyter-xarray-tiler.readthedocs.io/en/latest/contributor-guide/)!
+[Contributor Guide in our documentation](https://jupyter-geotiler.readthedocs.io/en/latest/contributor-guide/)!
