+++ docs/contributor-guide/index.md
@@ -7,6 +7,6 @@
 how-tos/index
 ```
 
-This documentation is aimed at contributors to `jupyter-xarray-tiler`.
+This documentation is aimed at contributors to `jupyter-geotiler`.
 
 👈 Use the navigation on the left to continue!
