+++ jupyter_geotiler/tests/test_xpublish_api.py
@@ -1,7 +1,7 @@
 import pytest
 from xarray import DataArray
 
-from jupyter_xarray_tiler.xpublish import (
+from jupyter_geotiler.xpublish import (
     _get_server,
     add_data_array,
     get_routes,
